Split ops module
